requirements.txt
matplotlib
numpy==1.26.4
pandas
lightning
ray[default,tune]==2.8.1
tune-sklearn
scienceplots
scikit_learn
scipy
torch
torchmetrics
torchvision
black
pvlib
tensorboard
tables
black[jupyter]
opencv-python
